+++ src/ui/units/ql/constants/index.ts
@@ -1,4 +1,4 @@
-import {QLChartType} from '../../../../shared';
+import {ConnectorType, QLChartType} from '../../../../shared';
 import {VISUALIZATION_IDS} from '../../../constants/visualizations';
 
 export enum AppStatus {
@@ -44,6 +44,32 @@ export const AVAILABLE_CHART_TYPES: QLChartType[] = [
     QLChartType.Promql,
     QLChartType.Monitoringql,
 ];
+
+// Connection types available for QL charts
+export const AVAILABLE_SQL_CONNECTION_TYPES = [
+    ConnectorType.ChFrozenDemo,
+    ConnectorType.ChOverYt,
+    ConnectorType.ChOverYtUserAuth,
+    ConnectorType.Chydb,
+    ConnectorType.Clickhouse,
+    ConnectorType.Greenplum,
+    ConnectorType.Mssql,
+    ConnectorType.Mysql,
+    ConnectorType.Oracle,
+    ConnectorType.Postgres,
+    ConnectorType.Ydb,
+    ConnectorType.Chyt,
+];
+
+export const AVAILABLE_PROMQL_CONNECTION_TYPES = ['promql'];
+
+export const AVAILABLE_MONITORINGQL_CONNECTION_TYPES = ['solomon', 'monitoring'];
+
+export const AVAILABLE_CONNECTION_TYPES_BY_CHART_TYPE = {
+    [QLChartType.Sql]: AVAILABLE_SQL_CONNECTION_TYPES,
+    [QLChartType.Promql]: AVAILABLE_PROMQL_CONNECTION_TYPES,
+    [QLChartType.Monitoringql]: AVAILABLE_MONITORINGQL_CONNECTION_TYPES,
+};
 
 export const QL_MOCKED_DATASET_ID = 'ql-mocked-dataset';
